Added a bears table and refactored to use routes
