+++ README.md
@@ -17,10 +17,10 @@ You'll write a server that lets you create and read _Zoos_ stored in a Relationa
 - [x] `cd` into your project folder.
 - [x] Run `npm install` or `yarn` to download the dependencies.
 - [x] Add `knex` and `sqlite3` npm modules.
-- Configure `knex` to connect to `/data/lambda.sqlite3` using the `sqlite3` module.
-- Write a set of endpionts inside `index.js` to satisfy the specifications listed below.
-- To start the API server, run `yarn start` or `npm start`.
-- Use _Postman_ to test your API.
+- [x] Configure `knex` to connect to `/data/lambda.sqlite3` using the `sqlite3` module.
+- [x] Write a set of endpionts inside `index.js` to satisfy the specifications listed below.
+- [x] To start the API server, run `yarn start` or `npm start`.
+- [x] Use _Postman_ to test your API.
 
 ## Specifications
